Fix an infinite loop of 6H-Dx on the first move.
If the AI plays 6H-Dx as its first move, that triggers Sx-Hx and
Sx-Hx. The second Sx-Hx puts it over the hand limit (6 cards) and
it can't figure out what to play. So we only solve for winning or
blocking moves when we're not chaining moves. This means the AI
won't always play perfectly. It might It might play 4H-Dx, and
then pick a card to discard that's not a block or win.